SCO.PaymentService update

+++ SCO.PaymentService.Domain/IPaymentLogic.cs
@@ -4,6 +4,6 @@ namespace SCO.PaymentService.Domain;
 
 public interface IPaymentLogic
 {
-    Task<PaymentResult> ProcessPayment(Guid orderId, double amount);
+    Task<PaymentResult> ProcessPayment(Guid orderId, decimal amount);
     Task<PaymentResult> AbortPayment(Guid paymentId);
 }
